Separate Test classes and address comments
