+++ src/test/java/edu/snu/mist/core/task/QueryManagerTest.java
@@ -26,8 +26,6 @@ import edu.snu.mist.common.operators.FlatMapOperator;
 import edu.snu.mist.common.operators.MapOperator;
 import edu.snu.mist.common.operators.ReduceByKeyOperator;
 import edu.snu.mist.common.rpc.RPCServerPort;
-import edu.snu.mist.common.sinks.Sink;
-import edu.snu.mist.common.sources.DataGenerator;
 import edu.snu.mist.common.sources.EventGenerator;
 import edu.snu.mist.common.sources.PunctuatedEventGenerator;
 import edu.snu.mist.common.types.Tuple2;
@@ -37,6 +35,8 @@ import edu.snu.mist.core.parameters.PlanStorePath;
 import edu.snu.mist.core.task.eventProcessors.parameters.DefaultNumEventProcessors;
 import edu.snu.mist.core.task.globalsched.parameters.GroupSchedModelType;
 import edu.snu.mist.core.task.stores.QueryInfoStore;
+import edu.snu.mist.core.task.utils.TestDataGenerator;
+import edu.snu.mist.core.task.utils.TestWithCountDownSink;
 import edu.snu.mist.formats.avro.AvroOperatorChainDag;
 import edu.snu.mist.formats.avro.Direction;
 import junit.framework.Assert;
@@ -48,12 +48,8 @@ import org.apache.reef.tang.Tang;
 import org.junit.Test;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
@@ -188,9 +184,9 @@ public final class QueryManagerTest {
     final List<String> sink1Result = new LinkedList<>();
     final List<Integer> sink2Result = new LinkedList<>();
     final PhysicalSink sink1 = new PhysicalSinkImpl<>("sink1",
-        null, new TestSink<String>(sink1Result, countDownAllOutputs));
+        null, new TestWithCountDownSink<String>(sink1Result, countDownAllOutputs));
     final PhysicalSink sink2 = new PhysicalSinkImpl<>("sink2",
-        null, new TestSink<Integer>(sink2Result, countDownAllOutputs));
+        null, new TestWithCountDownSink<Integer>(sink2Result, countDownAllOutputs));
 
     // Fake operator chain dag of QueryManager
     final AvroOperatorChainDag fakeOperatorChainDag = new AvroOperatorChainDag();
@@ -375,105 +371,6 @@ public final class QueryManagerTest {
     public static List<Map<String, Integer>> combine(final List<Map<String, Integer>> list1,
                                                      final List<Map<String, Integer>> list2) {
       return list1;
-    }
-  }
-
-  /**
-   * Test source generator which generates inputs from List.
-   */
-  final class TestDataGenerator<String> implements DataGenerator<String> {
-
-    private final AtomicBoolean closed;
-    private final AtomicBoolean started;
-    private final ExecutorService executorService;
-    private final long sleepTime;
-    private EventGenerator eventGenerator;
-    private final Iterator<String> inputs;
-
-    /**
-     * Generates input data from List and count down the number of input data.
-     */
-    TestDataGenerator(final List<String> inputs) {
-      this.executorService = Executors.newSingleThreadExecutor();
-      this.closed = new AtomicBoolean(false);
-      this.started = new AtomicBoolean(false);
-      this.sleepTime = 1000L;
-      this.inputs = inputs.iterator();
-    }
-
-    @Override
-    public void start() {
-      if (started.compareAndSet(false, true)) {
-        executorService.submit(() -> {
-          while (!closed.get()) {
-            try {
-              // fetch an input
-              final String input = nextInput();
-              if (eventGenerator == null) {
-                throw new RuntimeException("EventGenerator should be set in " +
-                    TestDataGenerator.class.getName());
-              }
-              if (input == null) {
-                Thread.sleep(sleepTime);
-              } else {
-                eventGenerator.emitData(input);
-              }
-            } catch (final IOException e) {
-              e.printStackTrace();
-              throw new RuntimeException(e);
-            } catch (final InterruptedException e) {
-              e.printStackTrace();
-            }
-          }
-        });
-      }
-    }
-
-    public String nextInput() throws IOException {
-      if (inputs.hasNext()) {
-        final String input = inputs.next();
-        return input;
-      } else {
-        return null;
-      }
-    }
-
-    @Override
-    public void close() throws Exception {
-      if (closed.compareAndSet(false, true)) {
-        executorService.shutdown();
-      }
-    }
-
-    @Override
-    public void setEventGenerator(final EventGenerator eventGenerator) {
-      this.eventGenerator = eventGenerator;
-    }
-  }
-
-  /**
-   * Test sink.
-   * It receives inputs, adds them to list, and countdown.
-   */
-  final class TestSink<I> implements Sink<I> {
-    private final List<I> result;
-    private final CountDownLatch countDownLatch;
-
-    TestSink(final List<I> result,
-             final CountDownLatch countDownLatch) {
-      this.result = result;
-      this.countDownLatch = countDownLatch;
-    }
-
-    @Override
-    public void close() throws Exception {
-      // do nothing
-    }
-
-    @Override
-    public void handle(final I input) {
-      result.add(input);
-      countDownLatch.countDown();
     }
   }
 }
